refactor units types folder location
