+++ src/components/editor/types/index.ts
@@ -1,3 +1,1 @@
-export * from './color';
-export * from './units';
 export * from './items';
