--- src/components/editor/types/units.ts
@@ -1,20 +1,0 @@
-export type Size =
-  | 'default'
-  | 'small'
-  | 'medium'
-  | 'large';
-
-export type Weight =
-  | 'default'
-  | 'light'
-  | 'regular'
-  | 'bold';
-
-export type Align =
-  | 'left'
-  | 'center'
-  | 'right';
-
-export type SpacingUnit = `${number}mm`;
-
-export type FontSizeUnit = `${number}pt`;
