+++ src/shared/types/units/index.ts
@@ -1,3 +1,7 @@
 export * from './alignment';
 export * from './size';
 export * from './weight';
+
+export type SpacingUnit = `${number}mm`;
+
+export type FontSizeUnit = `${number}pt`;
