+++ src/state/photo.ts
@@ -2,7 +2,7 @@ import { createSlice } from '@reduxjs/toolkit';
 import type { PayloadAction } from '@reduxjs/toolkit';
 import { HYDRATE } from 'next-redux-wrapper';
 
-import { Color } from '@/components/editor/types';
+import { Color } from '@/shared/types';
 import { AppState } from './store';
 
 export type PhotoState = {
