+++ stories/editor/common/inline.stories.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 
+import { Color } from '@/shared/types';
+
 import { Container } from '../../helpers/container';
 import { Box, Inline } from '../../../src/components/editor/common';
-import { Color } from '../../../src/components/editor/types';
 
 export default {
   title: 'editor/common/inline',
@@ -17,7 +18,7 @@ export default {
       defaultValue: 'blue',
       control: {
         type: 'select',
-        options: Color.values(),
+        options: Color.values,
       }
     },
   },
